add leave and cancel button
